On branch main  Your branch is ahead of 'origin/main' by 5 commits.    (use "git push" to publish your local commits)
 Changes to be committed:
	modified:   src/empresa_agro/Almacen.java
	modified:   src/empresa_agro/Producto.java

+++ src/empresa_agro/Producto.java
@@ -7,7 +7,7 @@ public abstract class Producto {
     private boolean isAlmacenado;
 
     private int stockEnKilos;
-    private double precioPorKilo;
+    private int precioPorKilo;
 
     public Producto(int diasCaducidad, String numLote) {
         
@@ -40,10 +40,10 @@ public abstract class Producto {
     public void setStockEnKilos(int stockEnKilos) {
         this.stockEnKilos = stockEnKilos;
     }
-    public double getPrecioPorKilo() {
+    public int getPrecioPorKilo() {
         return precioPorKilo;
     }
-    public void setPrecioPorKilo(double precioPorKilo) {
+    public void setPrecioPorKilo(int precioPorKilo) {
         this.precioPorKilo = precioPorKilo;
     }
